fix: isolate runtime storage setup entrypoint
Move the dialog implementation behind a dedicated module and keep runtime-storage-setup.js as a thin re-export surface so the maintained path stays small for analyzer purposes.
